Resolve the ambiguity by introducing an unambiguous syntax for type paremeters.
This proposal is not intended as a bikeshed on the specific syntax chosen,
but an alternative approach that removes the ambiguity present in the current
proposal that utilizes parenthesis for the new generic syntax. The choice of F@<T>
was pragmatic given that it could be added to the scanner in a few minutes.

See: https://gist.github.com/jboelter/5a7123a7723a72f88627f62ae405038d

See "Why not use the syntax `F(T)`?" for a detailed summary of the
ambiguities addressed by this proposal.